requirements.txt
amqp==5.1.1
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
billiard==4.1.0
boto3==1.28.32
botocore==1.31.32
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.4
cryptography==41.0.3
debugpy==1.6.7
decorator==5.1.1
Django==4.1.10
django-celery-results==2.5.1
django-redis==5.3.0
django-ses==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
dnspython==2.4.2
eventlet==0.33.3
executing==1.2.0
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
jmespath==1.0.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jwt==1.3.1
kombu==5.3.1
marshmallow==3.20.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pygame==2.4.0
Pygments==2.16.1
PyJWT==2.8.0
pymongo==3.13.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pyzmq==25.1.0
redis==5.0.0
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.2.4
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
